# orderworker/src/main/java/com/cristhianfdx/orderworker/model/Product.java
package com.cristhianfdx.orderworker.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class Product {
    private String productId;
    private String name;
    private double price;
}